+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Remove duplicate functions for image retrieval (`checkImage`, `resolveImageName`) in `resource_volume.go`
 ### Fixes
 - Remove `credentials` field from `ionoscloud_mongo_cluster` resource
+- Remove `credentials` field from `ionoscloud_pg_cluster` data source
+### Documentation
+- Update documentation for K8s cluster, nodepools and shares
 
 ## 6.4.12
 ### Features